add support for mpv as a player
For now, python-vlc is still required even if vlc is not intented to be used. Likewise, pympv is required even if mpv is not used. This will probably be the case for all supported players. However, it should be possible to have them installed without needed the players themselves installed.

Note: when this is released (soon), the documentation needs to be updated to reflect new dependency requirements.

fixes #14