change Database.episodes to not return tuples
Now that the episode object has an ep_id property, it makes more sense for the EpisodeMenu itself to grab the id from the episode rather than supplying it to all callers of Database.episodes() since others rarely need it.

Note that we still don't want the EpisodeMenu to simply store a list of Episode's.