Play rate resetting to default.
If the user changed the playback speed while listening, it was being
reset to the default when a new play command was being executed.
This could happen by either the combination of a pause/resume action
or by the act of moving to the next item in the queue.

This change preserves any ad-hoc speed changes that the user has made
so they will persevere through these play commands. The default play
rate is unchaged and will still be the default when castero next starts.

This change caused the test_queue_change_rate test to fail
with a change_rate not called error so I changed it to behave in a
similar mannor to test_queue_change_volume.